refactored overwrite unnamed nodes in gltf.
