+++ src/client/editor/ConflictHandler.js
@@ -1,3 +1,44 @@
+function nodesToTree(nodes) {
+  if (!nodes) {
+    return;
+  }
+
+  nodes.forEach((node, i) => {
+    if (!node.userData) {
+      createTreePath(node, 0, i, nodes);
+    }
+    if (node.userData._visited) {
+      return;
+    }
+    createTreePath(node, 0, i, nodes);
+  });
+
+  return nodes;
+}
+
+function createTreePath(node, layer, index, nodes) {
+  if (!node.userData) {
+    node.userData = {};
+  }
+  node.userData._visited = true;
+  if (node.userData._path) {
+    node.userData._path.push(index);
+  } else {
+    node.userData._path = [0, 0];
+  }
+
+  if (node.children) {
+    node.children.forEach((index, i) => {
+      const child = nodes[index];
+      if (!child.userData) {
+        child.userData = {};
+      }
+      child.userData._path = node.userData._path.slice(0);
+      createTreePath(child, layer + 1, i, nodes);
+    });
+  }
+}
+
 export default class ConflictHandler {
   constructor() {
     this._conflicts = {
@@ -125,6 +166,15 @@ export default class ConflictHandler {
       return;
     }
     return this._updatedNodes[hashPath];
+  };
+
+  updateNodeNames = nodes => {
+    const nodeTree = nodesToTree(nodes);
+
+    for (const node of nodeTree) {
+      node.name = this.getUpdatedNodeName(node.userData._path);
+      delete node.userData;
+    }
   };
 
   isUpdateNeeded = () => {
